refactor(notebook-cloud): move notebook viewer into a focused module
